* more methods in paver.path now check for existence or lack thereof   and won't fail as a result. (mkdir and makedirs both check that the   directory does not exist, rmdir and rmtree check to be sure that   it does.) This is because the goal is ultimately to create or remove   something... paver just makes sure that it either exists or doesn't.

+++ paver/path.py
@@ -902,13 +902,16 @@ class path(_base):
             dry("mkdir %s (mode %s)" % (self, mode), os.mkdir, self, mode)
 
     def makedirs(self, mode=0777):
-        dry("makedirs %s (mode %s)" % (self, mode), os.makedirs, self, mode)
+        if not self.exists():
+            dry("makedirs %s (mode %s)" % (self, mode), os.makedirs, self, mode)
 
     def rmdir(self):
-        dry("rmdir %s" % (self), os.rmdir, self)
+        if self.exists():
+            dry("rmdir %s" % (self), os.rmdir, self)
 
     def removedirs(self):
-        dry("removedirs %s" % (self), os.removedirs, self)
+        if self.exists():
+            dry("removedirs %s" % (self), os.removedirs, self)
 
 
     # --- Modifying operations on files
